phpcs violations corrected.  Some organizational changes done.
-phpcs violations corrected so that the build passes with the current
build process.
-Moved zillow housing data cron to crons folder and out of Utils
namespace.
-Removed the Utils namespace (which will break the crime cron)
