+++ src/HRQLS/Bootstrap.php
@@ -2,14 +2,14 @@
 /**
  * This is the first application file that gets loaded by the index.php file.
  *
- * @package JNT
+ * @package HRQLS
  */
 namespace HRQLS;
 
 use Silex\Application;
 use Silex\Provider\TwigServiceProvider;
 use Monolog\Logger;
-use JsonSchema\Uri\UriRetriever as UriRetriever;
+use JsonSchema\Uri\UriRetriever;
 
 /**
  * The class that manages startup of the application.
@@ -63,14 +63,22 @@ class Bootstrap
 
     /**
      * Array of URL's used by Crime Scraper
+     *
+     * @var array
      */
     private $crimeURLs = [
-        'chesapeake' => 'http://hamptonroads.com/newsdata/crime/chesapeake/search/rss?me=%2Fchesapeake%2Fsearch&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
-        'newportnews' => 'http://hamptonroads.com/newsdata/crime/newport-news/search/rss?me=%2Fnewport-news%2Fsearch&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
-        'norfolk' => 'http://hamptonroads.com/newsdata/crime/norfolk/search/rss?me=%2Fnorfolk%2Fsearch&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
-        'portsmouth' => 'http://hamptonroads.com/newsdata/crime/portsmouth/search/rss?me=%2Fportsmouth%2Fsearch&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
-        'suffolk' =>'http://hamptonroads.com/newsdata/crime/suffolk/search?me=%2Fsuffolk%2Fsearch&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
-        'vabeach' => 'http://hamptonroads.com/newsdata/crime/virginia-beach/search/rss?me=/virginia-beach/search&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
+        'chesapeake' => 'http://hamptonroads.com/newsdata/crime/chesapeake/search/rss?me=%2Fchesapeake%2Fsearch&' .
+            'type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
+        'newportnews' => 'http://hamptonroads.com/newsdata/crime/newport-news/search/rss?me=%2Fnewport-news%2Fsearch' .
+            '&type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
+        'norfolk' => 'http://hamptonroads.com/newsdata/crime/norfolk/search/rss?me=%2Fnorfolk%2Fsearch&type=&near=&' .
+            'radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
+        'portsmouth' => 'http://hamptonroads.com/newsdata/crime/portsmouth/search/rss?me=%2Fportsmouth%2Fsearch&' .
+            'type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
+        'suffolk' => 'http://hamptonroads.com/newsdata/crime/suffolk/search?me=%2Fsuffolk%2Fsearch&type=&near=&' .
+            'radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
+        'vabeach' => 'http://hamptonroads.com/newsdata/crime/virginia-beach/search/rss?me=/virginia-beach/search&' .
+            'type=&near=&radius=&op=Submit&form_token=9dc84572393ad9c68f54cad6549692f3&form_id=crime_searchform',
     ];
 
     /**
@@ -86,8 +94,7 @@ class Bootstrap
     public function __construct(Application $app)
     {
         $this->app = $app;
-        foreach ($this->crimeURLS as $city => $url)
-        {
+        foreach ($this->crimeURLS as $city => $url) {
             $objCrime = new CrimeConsumer($city, $url);
             $objCrime->consume();
         }
@@ -102,7 +109,7 @@ class Bootstrap
      */
     public function registerRoutes()
     {
-        foreach($this->config->routes->get as $endpoint) {
+        foreach ($this->config->routes->get as $endpoint) {
             $this->app->get($endpoint->url, $endpoint->controller);
         }
     }
@@ -124,7 +131,7 @@ class Bootstrap
 
         $this->config = [];
 
-        $this->config = json_decode( file_get_contents( __DIR__ . '/config/routes.json' ) );
+        $this->config = json_decode(file_get_contents(__DIR__ . '/config/routes.json'));
     }
 
     /**
